+++ kbe/src/lib/server/common.h
@@ -149,7 +149,7 @@ namespace KBEngine {
 			break;																							\
 		default:																							\
 			ERROR_MSG(fmt::format("not support componentType={}({})\n",										\
-					COMPONENTTYPE, COMPONENT_NAME_EX(COMPONENTTYPE)));										\
+					static_cast<int>(COMPONENTTYPE), COMPONENT_NAME_EX(COMPONENTTYPE)));										\
 			KBE_ASSERT(false && "not support componentType!\n");											\
 			break;																							\
 		};																									\
@@ -170,7 +170,7 @@ namespace KBEngine {
 			break;																							\
 		default:																							\
 			ERROR_MSG(fmt::format("not support componentType={}({})\n",										\
-					COMPONENTTYPE, COMPONENT_NAME_EX(COMPONENTTYPE)));										\
+					static_cast<int>(COMPONENTTYPE), COMPONENT_NAME_EX(COMPONENTTYPE)));										\
 			KBE_ASSERT(false && "not support componentType!\n");											\
 			break;																							\
 		};																									\
