+++ kbe/src/server/baseapp/baseapp.cpp
@@ -4583,7 +4583,7 @@ void Baseapp::onEntityCall(Network::Channel* pChannel, KBEngine::MemoryStream& s
 		default:
 			{
 				ERROR_MSG(fmt::format("Baseapp::onEntityCall: entitycallType {} error! must a baseType. entityID={}.\n",
-					calltype, eid));
+					static_cast<int>(calltype), eid));
 			}
 	};
